+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
+
 
 class CommentController extends Controller
 {
@@ -18,12 +20,15 @@ class CommentController extends Controller
         $comment->title = $request->title;
         $comment->text = $request->text;
         $comment->article_id = $request->article_id;
+        // $comment->author_id = Auth::id();
+        $comment->user()->associate(auth()->user());
         $comment->save();
         return redirect()->route('article.show', ['article'=>$comment->article_id]);
     }
 
     public function edit($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('comment', $comment);
         return view('comments.edit', ['comment'=>$comment]);
     }
 
@@ -42,6 +47,7 @@ class CommentController extends Controller
 
     public function delete($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('comment', $comment);
         $comment->delete();
         return redirect()->route('article.show', ['article'=>$comment->article_id]);
     }
