+++ resources/views/articles/show.blade.php
@@ -52,8 +52,10 @@
     <h5 class="card-title">{{$comment->title}}</h5>
     <h6 class="card-subtitle mb-2 text-body-secondary">{{$comment->text}}</h6>
     <div class="flex">
-    <a href="/comment/edit/{{$comment->id}}" class="btn btn-primary mr-1">Edit comment</a>
-    <a href="/comment/delete/{{$comment->id}}" class="btn btn-secondary">Delete comment</a>
+      @can('comment', $comment) 
+        <a href="/comment/edit/{{$comment->id}}" class="btn btn-primary mr-1">Edit comment</a>
+        <a href="/comment/delete/{{$comment->id}}" class="btn btn-secondary">Delete comment</a>
+      @endcan
 </div>
 </div>
 </div>
